ai_webagent/agent.py
from dotenv import load_dotenv
from prompts import AGENT_INSTRUCTION, SESSION_INSTRUCTION
from livekit import agents
from livekit.agents import AgentSession, Agent, RoomInputOptions, function_tool
from livekit.plugins import (
    google,
    noise_cancellation,
)
from mcp_client import MCPServerSse
from mcp_client.agent_tools import MCPToolsIntegration
import os
import logging
import signal
import sys
import asyncio
from tools import open_url
from kb import get_kb_answer

# Set up logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Global variable to track if shutdown is requested
shutdown_requested = False

def signal_handler(signum, frame):
    """Handle shutdown signals gracefully"""
    global shutdown_requested
    logger.info(f"Received signal {signum}. Initiating graceful shutdown...")
    shutdown_requested = True
    sys.exit(0)

# Register signal handlers
signal.signal(signal.SIGINT, signal_handler)
signal.signal(signal.SIGTERM, signal_handler)

load_dotenv()


@function_tool
async def answer_upclosets_question(question: str) -> str:
    """
    Answer questions about UpClosets of NOVA by referencing the knowledge base.
    
    Args:
        question: A question about UpClosets' services, products, processes, or information.
        
    Returns:
        A relevant answer based on UpClosets' knowledge base.
    """
    logger.info(f"answer_upclosets_question called with question: {question}")
    try:
        answer = get_kb_answer(question)
        logger.info(f"Knowledge base answer retrieved successfully. Answer length: {len(answer)}")
        return answer
    except Exception as e:
        logger.error(f"Error in answer_upclosets_question: {str(e)}")
        return f"I'm having trouble accessing the UpClosets information right now. Please try again later."


class Assistant(Agent):
    def __init__(self) -> None:
        logger.info("Initializing Assistant agent with tools: [open_url, answer_upclosets_question]")
        super().__init__(instructions=AGENT_INSTRUCTION,
                         tools=[open_url, answer_upclosets_question],)
        logger.info("Assistant agent initialized successfully")


async def entrypoint(ctx: agents.JobContext):
    global shutdown_requested
    
    logger.info("Starting agent entrypoint")
    logger.info(f"Job context: {ctx}")
    
    try:
        session = AgentSession(
            llm=google.beta.realtime.RealtimeModel(
                voice="Aoede",  # Female voice
            ),
        )
        logger.info("AgentSession created with Google Realtime Model (voice: Aoede)")

        mcp_server_url = os.environ.get("N8N_MCP_SERVER_URL")
        logger.info(f"MCP Server URL from environment: {mcp_server_url}")
        
        mcp_server = MCPServerSse(
            params={"url": mcp_server_url},
            cache_tools_list=True,
            name="SSE MCP Server"
        )
        logger.info("MCP Server created")

        logger.info("Creating agent with MCP tools integration")
        agent = await MCPToolsIntegration.create_agent_with_tools(
            agent_class=Assistant,
            mcp_servers=[mcp_server]
        )
        logger.info("Agent with MCP tools created successfully")

        logger.info("Starting session with room and agent")
        await session.start(
            room=ctx.room,
            agent=agent,
            room_input_options=RoomInputOptions(
                # LiveKit Cloud enhanced noise cancellation
                # - If self-hosting, omit this parameter
                # - For telephony applications, use `BVCTelephony` for best results
                noise_cancellation=noise_cancellation.BVC(),
            ),
        )
        logger.info("Session started successfully")

        logger.info("Connecting to context")
        await ctx.connect()
        logger.info("Connected to context successfully")

        logger.info("Generating initial reply with session instructions")
        await session.generate_reply(
            instructions=SESSION_INSTRUCTION,
        )
        logger.info("Initial reply generated successfully")
        
        # Keep the session alive until shutdown is requested
        while not shutdown_requested:
            await asyncio.sleep(1)
            
    except Exception as e:
        logger.error(f"Error in agent entrypoint: {e}")
        raise
    finally:
        logger.info("Agent entrypoint cleanup completed")


if __name__ == "__main__":
    agents.cli.run_app(agents.WorkerOptions(entrypoint_fnc=entrypoint))
